IFS-4421 : Align with grant send JSON contract

+++ ifs-sil-stub/src/test/java/org/innovateuk/ifs/sil/grant/controller/GrantEndpointControllerMockMvcTest.java
@@ -57,7 +57,7 @@ public class GrantEndpointControllerMockMvcTest extends AbstractEndpointControll
 
     private Set<Period> createPeriods(int durationInMonths, BigDecimal value) {
         return Stream.iterate(0, i -> i + 1).limit(durationInMonths)
-                .map(i -> new Period().month(i + 1).value(value))
+                .map(i -> new Period().month(i + 1).value(value.longValue()))
                 .collect(Collectors.toSet());
     }
